+++ xdr/account_id_test.go
@@ -20,3 +20,19 @@ var _ = Describe("xdr.AccountId#Address()", func() {
 		Expect(addy).To(Equal("GCR22L3WS7TP72S4Z27YTO6JIQYDJK2KLS2TQNHK6Y7XYPA3AGT3X4FH"))
 	})
 })
+
+var _ = Describe("xdr.AccountId#Equals()", func() {
+	It("returns true when the account ids have equivalent values", func() {
+		var l, r AccountId
+		l.SetAddress("GCR22L3WS7TP72S4Z27YTO6JIQYDJK2KLS2TQNHK6Y7XYPA3AGT3X4FH")
+		r.SetAddress("GCR22L3WS7TP72S4Z27YTO6JIQYDJK2KLS2TQNHK6Y7XYPA3AGT3X4FH")
+		Expect(l.Equals(r)).To(BeTrue())
+	})
+
+	It("returns false when the account ids have different values", func() {
+		var l, r AccountId
+		l.SetAddress("GCR22L3WS7TP72S4Z27YTO6JIQYDJK2KLS2TQNHK6Y7XYPA3AGT3X4FH")
+		r.SetAddress("GBTBXQEVDNVUEESCTPUT3CHJDVNG44EMPMBELH5F7H3YPHXPZXOTEWB4")
+		Expect(l.Equals(r)).To(BeFalse())
+	})
+})