+++ CHANGELOG.md
@@ -9,6 +9,8 @@ bumps.  A breaking change will get clearly notified in this log.
 ## [Unreleased]
 
 ### Added
+- `*xdr.AccountId` learned `Equals(AccountId)`, helper to compare account ids for equality.
+- The `amount` package learned `MustParse(string)`.
 - `*xdr.Price` learned `String()`.
 - `*xdr.PathPaymentResult` learned `SendAmount()`, a helper to extract how much of the source asset was spent when making a multi-asset payment.